refractored midichannel to its own class and moved playIntro to main
