+++ src/piano/Main.java
@@ -1,16 +1,42 @@
 package piano;
 
+import javax.sound.midi.MidiChannel;
 import javax.sound.midi.MidiUnavailableException;
 
 public class Main {
     public static void main(String[] args) {
         try {
-            PianoGUI frame = new PianoGUI();
+            MidiChannel midiChannel = new Channel().getChannel();
+            PianoGUI frame = new PianoGUI(midiChannel);
             frame.setVisible(true);
-            frame.playIntro();
-        } catch (MidiUnavailableException exc) {
-            exc.printStackTrace();
+            playIntro(midiChannel);
+        } catch (MidiUnavailableException e) {
+            e.printStackTrace();
         }
+    }
 
+    private static void playIntro(MidiChannel channel)
+    {
+        try
+        {
+            int[] notes = { Notes.C, Notes.D, Notes.E, Notes.F, Notes.G, Notes.A, Notes.B };
+            for (int i = 0; i < notes.length; ++i)
+            {
+                channel.noteOn(notes[i], SoundSettings.VOLUME);
+                Thread.sleep(100);
+                channel.noteOff(notes[i]);
+            }
+            // Play a C major chord.
+            channel.noteOn(Notes.C, SoundSettings.VOLUME);
+            channel.noteOn(Notes.E, SoundSettings.VOLUME);
+            channel.noteOn(Notes.G, SoundSettings.VOLUME);
+            Thread.sleep(3000);
+            channel.allNotesOff();
+            Thread.sleep(500);
+        }
+        catch (InterruptedException e)
+        {
+            e.printStackTrace();
+        }
     }
 }
